+++ projects/js-packages/publicize-components/src/components/unified-modal/social-post-preview/use-modal-screen.tsx
@@ -1,6 +1,8 @@
+import { useBreakpoint } from '@automattic/viewport-react';
 import { __, sprintf, _n } from '@wordpress/i18n';
-import { useMemo, useState } from 'react';
+import { useId, useMemo, useState } from 'react';
 import useSocialMediaConnections from '../../../hooks/use-social-media-connections';
+import { Connection } from '../../../social-store/types';
 import { ScreenDetails } from '../types';
 import { Content } from './content';
 import { Sidebar } from './sidebar';
@@ -13,20 +15,30 @@ import { Sidebar } from './sidebar';
 export function useModalScreen(): ScreenDetails {
 	const { connections, enabledConnections } = useSocialMediaConnections();
 
-	const [ selectedConnection, setSelectedConnection ] = useState( connections[ 0 ] );
+	const [ selectedConnection, setSelectedConnection ] = useState< Connection >( connections[ 0 ] );
+
+	const baseId = useId();
+	const isSmallScreen = useBreakpoint( '<660px' );
 
 	return useMemo(
 		() => ( {
 			path: '/',
-			title: __( 'Customize and preview social posts', 'jetpack-publicize-components' ),
+			title: __( 'Preview and customize', 'jetpack-publicize-components' ),
 			isScreenLocked: true,
-			sidebar: (
+			sidebar: isSmallScreen ? null : (
 				<Sidebar
-					onClickConnection={ setSelectedConnection }
+					baseId={ baseId }
+					onSelectConnection={ setSelectedConnection }
 					selectedConnection={ selectedConnection }
 				/>
 			),
-			content: <Content selectedConnection={ selectedConnection } />,
+			content: (
+				<Content
+					baseId={ baseId }
+					selectedConnection={ selectedConnection }
+					forSmallScreen={ isSmallScreen }
+				/>
+			),
 			footerContent: enabledConnections.length ? (
 				<span>
 					{ sprintf(
@@ -49,6 +61,6 @@ export function useModalScreen(): ScreenDetails {
 				},
 			],
 		} ),
-		[ selectedConnection, setSelectedConnection, enabledConnections ]
+		[ isSmallScreen, baseId, selectedConnection, enabledConnections.length ]
 	);
 }