+++ projects/packages/connection/tests/php/XMLRPC_Connector_Test.php
@@ -59,7 +59,7 @@ class XMLRPC_Connector_Test extends TestCase {
 		openssl_sign(
 			wp_json_encode(
 				$url_parameters,
-				0 // No `json_encode()` flags because this needs to match whatever is calculating the hash on the other end.
+				0 // phpcs:ignore Jetpack.Functions.JsonEncodeFlags.ZeroFound -- No `json_encode()` flags because this needs to match whatever is calculating the hash on the other end.
 			),
 			$signature,
 			$private_key