+++ projects/plugins/jetpack/modules/woocommerce-analytics/classes/class-jetpack-woocommerce-analytics-universal.php
@@ -95,8 +95,14 @@ class Jetpack_WooCommerce_Analytics_Universal {
 	 * @deprecated 13.3
 	 */
 	public function remove_from_cart() {
-		$common_props = $this->render_properties_as_js(
-			$this->get_common_properties()
+		$common_props = wp_json_encode(
+			array_merge(
+				array(
+					'_en' => 'woocommerceanalytics_remove_from_cart',
+				),
+				$this->get_common_properties()
+			),
+			JSON_UNESCAPED_SLASHES | JSON_HEX_TAG | JSON_HEX_AMP
 		);
 
 		// We listen at div.woocommerce because the cart 'form' contents get forcibly
@@ -106,17 +112,11 @@ class Jetpack_WooCommerce_Analytics_Universal {
 			"jQuery( 'div.woocommerce' ).on( 'click', 'a.remove', function() {
 				var productID = jQuery( this ).data( 'product_id' );
 				var quantity = jQuery( this ).parent().parent().find( '.qty' ).val()
-				var productDetails = {
-					'id': productID,
-					'quantity': quantity ? quantity : '1',
-				};
-				_wca.push( {
-					'_en': 'woocommerceanalytics_remove_from_cart',
-					'pi': productDetails.id,
-					'pq': productDetails.quantity, " .
-					$common_props . '
-				} );
-			} );'
+				var common_props = $common_props;
+				common_props.pi = productID;
+				common_props.pq = quantity ? quantity : '1';
+				_wca.push( common_props );
+			} );"
 		);
 	}
 
@@ -402,8 +402,14 @@ class Jetpack_WooCommerce_Analytics_Universal {
 	 * @deprecated 13.3
 	 */
 	public function remove_from_cart_via_quantity() {
-		$common_props = $this->render_properties_as_js(
-			$this->get_common_properties()
+		$common_props = wp_json_encode(
+			array_merge(
+				array(
+					'_en' => 'woocommerceanalytics_remove_from_cart',
+				),
+				$this->get_common_properties()
+			),
+			JSON_UNESCAPED_SLASHES | JSON_HEX_TAG | JSON_HEX_AMP
 		);
 
 		wc_enqueue_js(
@@ -414,14 +420,12 @@ class Jetpack_WooCommerce_Analytics_Universal {
 					var qty = jQuery( this ).find( 'input.qty' );
 					if ( qty && qty.val() === '0' ) {
 						var productID = jQuery( this ).find( '.product-remove a' ).data( 'product_id' );
-						_wca.push( {
-							'_en': 'woocommerceanalytics_remove_from_cart',
-							'pi': productID, " .
-							$common_props . '
-						} );
+						var common_props = $common_props;
+						common_props.pi = productID;
+						_wca.push( common_props );
 					}
 				} );
-			} );'
+			} );"
 		);
 	}