+++ projects/packages/backup/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.33] - 2025-12-08
+### Fixed
+- Ensure proper flags are used with `json_encode()`. [#46092]
+
 ## [4.2.32] - 2025-12-01
 ### Changed
 - Update package dependencies. [#46143]
@@ -990,6 +994,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add API endpoints and Jetpack Backup package for managing Help…
 
+[4.2.33]: https://github.com/Automattic/jetpack-backup/compare/v4.2.32...v4.2.33
 [4.2.32]: https://github.com/Automattic/jetpack-backup/compare/v4.2.31...v4.2.32
 [4.2.31]: https://github.com/Automattic/jetpack-backup/compare/v4.2.30...v4.2.31
 [4.2.30]: https://github.com/Automattic/jetpack-backup/compare/v4.2.29...v4.2.30